requirements.txt
pandas==2.1.4
gspread==5.12.0
oauth2client==4.1.3
beautifulsoup4==4.12.2
requests==2.31.0
scikit-learn==1.3.2
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
aiohttp==3.9.1
selenium==4.16.0
webdriver-manager==4.0.1
nltk==3.8.1
affinity-api==1.0.0 